+++ src/scripts.js
@@ -2,6 +2,12 @@ import {
   getData
 } from './apiCalls.js'
 
+import {
+  createSleepChart,
+  createHydrationChart,
+  createCompareDonut
+} from './graphs.js'
+
 import './css/styles.css';
 import './images/turing-logo.png'
 
@@ -21,7 +27,8 @@ const dateSelected = document.querySelector(".date-selection");
 const buttons= Array.from(document.querySelectorAll(".dataButton"));
 const widgets = Array.from(document.querySelectorAll(".widget"));
 const weeklyHydration = document.querySelector("#weeklyHydrationChart");
-const weeklySleep = document.querySelector("#weeklySleepChart")
+const weeklySleep = document.querySelector("#weeklySleepChart");
+const stepComparisonDonut = document.querySelector("#stepComparisonDonut");
 
 //global variables
 let userRepo;
@@ -46,13 +53,10 @@ function fetchUsers() {
     activityData = data[3].activityData;
     createUserRepo();
     displayUserDetails();
-    compareSteps();
     displaySnapshotData();
-    // displayTodaysWaterIntake();
+    compareSteps();
     displayWeeklyWaterIntake();
-    // displayTodaysSleepStats();
     displayWeeklySleep();
-    // displayAverageSleepData();
   })
 }
 
@@ -66,13 +70,6 @@ function displayUserDetails(){
   Welcome! ${currentUser.name}
   ${currentUser.address}
   ${currentUser.email}`;
-}
-
-function compareSteps() {
-  activityWidget.innerText = `Hey ${currentUser.returnFirstName()}!
-  This is how your step goal compares to other users!
-  Yours: ${currentUser.dailyStepGoal} vs Theirs: ${userRepo.getAverageStepGoal()}`
-
 }
 
 function displayTodaysWaterIntake() {
@@ -94,51 +91,20 @@ function displaySnapshotData() {
   `
 }
 
+function compareSteps() {
+  const config = createCompareDonut(currentUser, userRepo.getAverageStepGoal());
+  const stepComparison = new Chart(stepComparisonDonut, config)
+  // activityWidget.innerText = `Hey ${currentUser.returnFirstName()}!
+  // This is how your step goal compares to other users!
+  // Yours: ${currentUser.dailyStepGoal} vs Theirs: ${userRepo.getAverageStepGoal()}`
+
+}
+
 function displayWeeklyWaterIntake() {
   //refactor date to be dynamic
   let weeklyWaterIntake = userRepo.getDailyFluidIntakeByWeek(11, "2020/01/22")
   const config = createHydrationChart(weeklyWaterIntake);
   const weeklyHydrationChart = new Chart(weeklyHydration, config)
-}
-
-function createHydrationChart(weeklyWaterIntake) {
-  const labels = weeklyWaterIntake.map(intake => intake.date);
-  const values = weeklyWaterIntake.map(intake => intake.fluidOz)
-  const data = {
-    labels: labels,
-    datasets: [{
-      label: "Weekly Hydration",
-      backgroundColor: [
-        'rgba(255, 99, 132, 0.2)',
-        'rgba(255, 159, 64, 0.2)',
-        'rgba(255, 205, 86, 0.2)',
-        'rgba(75, 192, 192, 0.2)',
-        'rgba(54, 162, 235, 0.2)',
-        'rgba(153, 102, 255, 0.2)',
-        'rgba(201, 203, 207, 0.2)'
-      ],
-      borderColor: [
-        'rgb(255, 99, 132)',
-        'rgb(255, 159, 64)',
-        'rgb(255, 205, 86)',
-        'rgb(75, 192, 192)',
-        'rgb(54, 162, 235)',
-        'rgb(153, 102, 255)',
-        'rgb(201, 203, 207)'
-      ],
-      borderWidth: 1,
-      data: values
-    }]
-  };
-
-  const config = {
-    type: "bar",
-    data: data,
-    options: {
-      barThickness: 20,
-    }
-  }
-  return config
 }
 
 function displayTodaysSleepStats() {
@@ -153,53 +119,7 @@ function displayWeeklySleep() {
   let quality = userRepo.getQualityByWeek(11, "2020/01/22");
   const config = createSleepChart(sleep, quality)
   return new Chart(weeklySleep, config)
-  // sleepWidget.innerHTML += `<br>Your Weekly sleep stats:<br>`
-  // weeklySleep.forEach((sleepData, index) => {
-  //   sleepWidget.innerHTML += `${sleepData.date}: Hours: ${sleepData.hoursSlept}
-  //   Quality: ${weeklyQuality[index].sleepQuality}<br>
-  //   `
-  // })
 };
-function createSleepChart(weeklySleep, weeklyQuality) {
-  const labels = weeklySleep.map(sleep => sleep.date);
-  const sleepValues = weeklySleep.map(sleep => sleep.hoursSlept)
-  const qualityValues = weeklyQuality.map(quality => quality.sleepQuality)
-
-  const data = {
-    labels: labels,
-    datasets: [{
-      label: "Weekly Sleep Hours",
-      backgroundColor: [
-        'rgba(255, 99, 132, 0.2)'
-      ],
-      borderColor: [
-        'rgb(255, 99, 132)'
-      ],
-      borderWidth: 1,
-      data: sleepValues
-    },
-    {
-      label: "Weekly Sleep Quality",
-      backgroundColor: [
-        'rgba(201, 203, 207, 0.2)'
-      ],
-      borderColor: [
-        'rgb(201, 203, 207)'
-      ],
-      borderWidth: 1,
-      data: qualityValues
-    }]
-  };
-
-  const config = {
-    type: "bar",
-    data: data,
-    options: {
-      barThickness: 20,
-    }
-  }
-  return config
-}
 
 function displayAverageSleepData() {
   let allTimeSleep = userRepo.getAverageSleep(11)
